chore: move project layout to one single application
