--- go.work
@@ -1,7 +1,0 @@
-go 1.19
-
-use (
-	./cmd/api
-	./cmd/ratings
-	./cmd/talks
-)
