+++ internal/ratings/types.go
