+++ internal/talks/types.go
